perf: settings ux repeat render
