+++ docSite/content/docs/installation/upgrading/45.md
@@ -72,10 +72,11 @@ CREATE INDEX CONCURRENTLY vector_index ON modeldata USING hnsw (vector vector_ip
 
 ### Fast GPT V4.5
 
-1. 升级 PgVector 插件，引入 HNSW 索引。
-2. 优化 TextSplitter
-3. AI对话模块，增加【返回AI内容】选项，可控制 AI 的内容不直接返回浏览器。
-4. 修复 - 分享链接鉴权问题
+1. 新增 - 升级 PgVector 插件，引入 HNSW 索引，极大加快的知识库搜索速度。
+2. 新增 - AI对话模块，增加【返回AI内容】选项，可控制 AI 的内容不直接返回浏览器。
+3. 优化 - TextSplitter，采用递归拆解法。
+4. 优化 - 高级编排 UX 性能
+5. 修复 - 分享链接鉴权问题
