+++ projects/app/public/docs/versionIntro.md
@@ -1,9 +1,10 @@
 ### Fast GPT V4.5.0
 
-1. 升级 PgVector 插件，引入 HNSW 索引，极大加快的知识库搜索速度。
-2. 优化 TextSplitter，采用递归拆解法。
-3. AI对话模块，增加【返回AI内容】选项，可控制 AI 的内容不直接返回浏览器。
-4. 优化数据集管理，区分手动录入和标注，可追数据至某个文件，保留链接读取的原始链接。
-5. [使用文档](https://doc.fastgpt.run/docs/intro/)
-6. [点击查看高级编排介绍文档](https://doc.fastgpt.run/docs/workflow)
-7. [点击查看商业版](https://doc.fastgpt.run/docs/commercial/)
+1. 新增 - 升级 PgVector 插件，引入 HNSW 索引，极大加快的知识库搜索速度。
+2. 新增 - AI对话模块，增加【返回AI内容】选项，可控制 AI 的内容不直接返回浏览器。
+3. 优化 - TextSplitter，采用递归拆解法。
+4. 优化 - 高级编排 UX 性能
+5. 优化数据集管理，区分手动录入和标注，可追数据至某个文件，保留链接读取的原始链接。
+6. [使用文档](https://doc.fastgpt.run/docs/intro/)
+7. [点击查看高级编排介绍文档](https://doc.fastgpt.run/docs/workflow)
+8. [点击查看商业版](https://doc.fastgpt.run/docs/commercial/)
